Add groupId to RabbitMQ model and integrate message consumption in WebSocketController
